@@ -1,8 +1,8 @@
-export { use, Instruction } from './instruction/use';
+export { use, type Instruction } from './instruction/use';
 export { get } from './instruction/get';
 export { set } from './instruction/set';
 export { ref } from './instruction/ref';
 
 export { State, State as default, METHOD } from './state';
-export { watch, addListener, event, Observable } from './control';
+export { watch, addListener, event, Observable } from './observable';
 export { Context } from './context';
