@@ -3,7 +3,7 @@
 import { vi, expect, it, describe, act, render, screen } from '../vitest';
 
 import React, { Children, Component, isValidElement } from 'react';
-import { Consumer, get, State, set } from '.';
+import State, { Consumer, get, set } from '.';
 
 it('will create and provide instance', () => {
   class Control extends State {
@@ -57,7 +57,7 @@ describe('new method', () => {
     const didCreate = vi.fn();
 
     class Test extends State {
-      new() {
+      protected new() {
         didCreate();
       }
     }
@@ -69,17 +69,6 @@ describe('new method', () => {
     element.rerender(<Test />);
 
     expect(didCreate).toHaveBeenCalledTimes(1);
-  });
-
-  it('will enforce signature', () => {
-    class Test extends State {
-      new(foo: string) {}
-    }
-
-    void function test() {
-      // @ts-expect-error
-      void (<Test />);
-    };
   });
 });
 
@@ -114,7 +103,7 @@ describe('element props', () => {
 
     render(<Foo value="barfoo" />);
 
-    expect(didSet).toBeCalled();
+    expect(didSet).toHaveBeenCalled();
   });
 
   it('will override method', async () => {
