@@ -1,14 +1,22 @@
-import './state.get';
-import './state.use';
+import { State } from '@expressive/mvc';
+import { SolidState } from './state';
 
-export {
-  State,
-  State as default,
-  get,
-  use,
-  ref,
-  set,
-  has
-} from '@expressive/mvc';
+/**
+ * Augmented State namespace to include SolidState hooks as static methods.
+ *
+ * This retroactively adds Solid support to agnostic State defined
+ * by component libraries which import `@expressive/mvc` directly.
+ */
+declare module '@expressive/mvc' {
+  namespace State {
+    export import get = SolidState.get;
+    export import use = SolidState.use;
+  }
+}
+
+State.get = SolidState.get;
+State.use = SolidState.use;
 
 export { Provider } from './context';
+export { SolidState as State, SolidState as default };
+export { Context, get, use, ref, set } from '@expressive/mvc';
